requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0

# API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
requests>=2.31.0

# Data Processing
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.0